further updates
added in the ability to create specific tags for numbers or urls. also
added in the ability to remove all tags and comments from files.

changed how the file list is retreived. it is now an iterator that
returns a result of trying to load the file from the list of paths
provided by the user.