tags update and dump
updated the code for tags since there is only one spot where the tags
data is set it is all in there now. removed all the add commands for
tags since it was a lot of duplicate logic. to replace all existing tags
there is a new command called "--replace".

added in the dump command that will do as it says and just dump the
contents of the database to stdout