collections added
went through and added the logic for creating collections of files.
currently working on updating the open command to work with a collection
but it needs to be fixed.

also went through and updated all the commands to use newer functions
for getting a db file and do only operate on a single db file that is
currently based on the current working directory.