path and small updates
updated functions to use a new file path iterator. it takes the list of
files and will create a full path and a relative path to the db
specified. also added in a new function to log an error result and
return an option with the value if it was an ok result.

removed the file src/file.rs since it was no longer needed with the
addition of the above.

also made some small changes in other places but functionality should
remain the same.