db keys change
changed the db keys to use strings vs system specific paths. previously
the paths were used and a system could have different separators which
would cause issues in systems that have a different separator than what
the key was originally made on. now it will check to make sure that the
path separator is '/' and if not then change it to that character.

also removed the coll update command as it does not do much and there is
another command for adding/removing files from a collection